+++ Sources/BFKit/Apple/UIKit/UIPasteboard+Extensions.swift
@@ -36,14 +36,14 @@ public extension UIPasteboard {
     /// Copy a text to the pasteboard.
     ///
     /// - Parameter text: The text to be copy to.
-    public static func copy(text: String) {
+    static func copy(text: String) {
         UIPasteboard.general.string = text
     }
     
     /// Returns the last copied string from pasteboard.
     ///
     /// - Returns: Returns the last copied string from pasteboard.
-    public static func getString() -> String {
+    static func getString() -> String {
         guard let pasteboard = UIPasteboard.general.string else {
             return ""
         }