+++ Sources/BFKit/Linux/BFKit/BFKit.swift
@@ -32,7 +32,7 @@ import Foundation
 #endif
 
 /// BFKit version string.
-public let BFKitVersion = "4.1.0"
+public let BFKitVersion = "5.0.0"
 
 /// BFKit author string.
 public let BFKitAuthor = "Fabrizio Brancati"