+++ Sources/BFKit/Linux/Foundation/FileManager+Extensions.swift
@@ -33,7 +33,7 @@ public extension FileManager {
     // MARK: - Variables
     
     /// Path type enum.
-    public enum PathType: Int {
+    enum PathType: Int {
         /// Main bundle path.
         case mainBundle
         /// Library path.
@@ -54,7 +54,7 @@ public extension FileManager {
     ///
     /// - Parameter path: Path type.
     /// - Returns: Returns the path type String.
-    public func pathFor(_ path: PathType) -> String? {
+    func pathFor(_ path: PathType) -> String? {
         var pathString: String?
         
         switch path {
@@ -87,7 +87,7 @@ public extension FileManager {
     ///   - path: File path.
     ///   - content: Content to be saved.
     /// - Throws: write(toFile:, atomically:, encoding:) errors.
-    public func save(file: String, in path: PathType, content: String) throws {
+    func save(file: String, in path: PathType, content: String) throws {
         guard let path = FileManager.default.pathFor(path) else {
             return
         }
@@ -102,7 +102,7 @@ public extension FileManager {
     ///   - path: File path.
     /// - Returns: Returns the content of the file a String.
     /// - Throws: Throws String(contentsOfFile:, encoding:) errors.
-    public func read(file: String, from path: PathType) throws -> String? {
+    func read(file: String, from path: PathType) throws -> String? {
         guard let path = FileManager.default.pathFor(path) else {
             return nil
         }
@@ -118,7 +118,7 @@ public extension FileManager {
     ///   - filename: PLIST filename.
     /// - Returns: Returns true if the operation was successful, otherwise false.
     @discardableResult
-    public func savePlist(object: Any, in path: PathType, filename: String) -> Bool {
+    func savePlist(object: Any, in path: PathType, filename: String) -> Bool {
         let path = checkPlist(path: path, filename: filename)
         
         guard !path.exist else {
@@ -134,7 +134,7 @@ public extension FileManager {
     ///   - path: Path of PLIST.
     ///   - filename: PLIST filename.
     /// - Returns: Returns the loaded object.
-    public func readPlist(from path: PathType, filename: String) -> Any? {
+    func readPlist(from path: PathType, filename: String) -> Any? {
         let path = checkPlist(path: path, filename: filename)
         
         guard !path.exist else {
@@ -162,7 +162,7 @@ public extension FileManager {
     ///
     /// - Parameter file: Filename.
     /// - Returns: Returns the path as a String.
-    public func mainBundlePath(file: String = "") -> String? {
+    func mainBundlePath(file: String = "") -> String? {
         return file.isEmpty ? Bundle.main.bundlePath : Bundle.main.path(forResource: file.deletingPathExtension, ofType: file.pathExtension)
     }
     
@@ -170,7 +170,7 @@ public extension FileManager {
     ///
     /// - Parameter file: Filename
     /// - Returns: Returns the path as a String.
-    public func documentsPath(file: String = "") -> String? {
+    func documentsPath(file: String = "") -> String? {
         guard let documentsURL = FileManager.default.urls(for: .documentDirectory, in: .userDomainMask).first else {
             return nil
         }
@@ -182,7 +182,7 @@ public extension FileManager {
     ///
     /// - Parameter file: Filename
     /// - Returns: Returns the path as a String.
-    public func libraryPath(file: String = "") -> String? {
+    func libraryPath(file: String = "") -> String? {
         guard let libraryURL = FileManager.default.urls(for: .libraryDirectory, in: .userDomainMask).first else {
             return nil
         }
@@ -194,7 +194,7 @@ public extension FileManager {
     ///
     /// - Parameter file: Filename
     /// - Returns: Returns the path as a String.
-    public func cachePath(file: String = "") -> String? {
+    func cachePath(file: String = "") -> String? {
         guard let cacheURL = FileManager.default.urls(for: .cachesDirectory, in: .userDomainMask).first else {
             return nil
         }
@@ -206,7 +206,7 @@ public extension FileManager {
     ///
     /// - Parameter file: Filename
     /// - Returns: Returns the path as a String.
-    public func applicationSupportPath(file: String = "") -> String? {
+    func applicationSupportPath(file: String = "") -> String? {
         guard let applicationSupportURL = FileManager.default.urls(for: .applicationSupportDirectory, in: .userDomainMask).first else {
             return nil
         }
@@ -226,7 +226,7 @@ public extension FileManager {
     ///
     /// - Parameter file: Filename.
     /// - Returns: Returns the path as a String.
-    public func temporaryPath(file: String = "") -> String? {
+    func temporaryPath(file: String = "") -> String? {
         return NSTemporaryDirectory().appendingPathComponent(file)
     }
     
@@ -237,7 +237,7 @@ public extension FileManager {
     ///   - path: Path of the file.
     /// - Returns: Returns the file size.
     /// - Throws: Throws FileManager.default.attributesOfItem(atPath:) errors.
-    public func size(file: String, from path: PathType) throws -> Float? {
+    func size(file: String, from path: PathType) throws -> Float? {
         if !file.isEmpty {
             guard let path = FileManager.default.pathFor(path) else {
                 return nil
@@ -260,7 +260,7 @@ public extension FileManager {
     ///   - file: File to delete.
     ///   - path: Path of the file.
     /// - Throws: Throws FileManager.default.removeItem(atPath:) errors.
-    public func delete(file: String, from path: PathType) throws {
+    func delete(file: String, from path: PathType) throws {
         if !file.isEmpty {
             guard let path = FileManager.default.pathFor(path) else {
                 throw BFKitError.pathNotExist
@@ -280,7 +280,7 @@ public extension FileManager {
     ///   - destination:  Destination path of the file.
     /// - Returns: Returns true if the operation was successful, otherwise false.
     /// - Throws: Throws FileManager.default.moveItem(atPath:, toPath:) and BFKitError errors.
-    public func move(file: String, from origin: PathType, to destination: PathType) throws {
+    func move(file: String, from origin: PathType, to destination: PathType) throws {
         let paths = try check(file: file, origin: origin, destination: destination)
         
         if paths.fileExist {
@@ -296,7 +296,7 @@ public extension FileManager {
     ///   - destination: Destination path
     /// - Returns: Returns true if the operation was successful, otherwise false.
     /// - Throws: Throws FileManager.default.copyItem(atPath:, toPath:) and BFKitError errors.
-    public func copy(file: String, from origin: PathType, to destination: PathType) throws {
+    func copy(file: String, from origin: PathType, to destination: PathType) throws {
         let paths = try check(file: file, origin: origin, destination: destination)
         
         if paths.fileExist {
@@ -337,7 +337,7 @@ public extension FileManager {
     ///   - newName: New filename.
     /// - Returns: Returns true if the operation was successful, otherwise false.
     /// - Throws: Throws FileManager.default.copyItem(atPath:, toPath:), FileManager.default.removeItem(atPath:, toPath:) and BFKitError errors.
-    public func rename(file: String, in origin: PathType, to newName: String) throws {
+    func rename(file: String, in origin: PathType, to newName: String) throws {
         guard let originPath = FileManager.default.pathFor(origin) else {
             throw BFKitError.pathNotExist
         }
@@ -360,7 +360,7 @@ public extension FileManager {
     /// - Returns: Returns true if the operation was successful, otherwise false.
     /// - Throws: Throws BFKitError errors.
     @discardableResult
-    public func setSettings(filename: String, object: Any, forKey objectKey: String) -> Bool {
+    func setSettings(filename: String, object: Any, forKey objectKey: String) -> Bool {
         guard var path = FileManager.default.pathFor(.applicationSupport) else {
             return false
         }
@@ -393,7 +393,7 @@ public extension FileManager {
     ///   - filename: Settings filename. "-Settings" will be automatically added.
     ///   - forKey: Object key.
     /// - Returns: Returns the object for the given key.
-    public func getSettings(filename: String, forKey objectKey: String) -> Any? {
+    func getSettings(filename: String, forKey objectKey: String) -> Any? {
         guard var path = FileManager.default.pathFor(.applicationSupport) else {
             return nil
         }